added new template list component for read only view
